% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAltModeTrips.R
\docType{data}
\name{AltModeModels_ls}
\alias{AltModeModels_ls}
\title{Alternative mode trip models}
\format{A list having 'Metro' and 'NonMetro' components. Each component has
the following components:
Walk: a hurdle model of walk trips having Count and Zero components;
Bike: a hurdle model of bike trips having Count and Zero components;
Transit: a hurdle model of public transit trips having Count and Zero
components;}
\source{
CalculateAltModeTrips.R.
}
\usage{
AltModeModels_ls
}
\description{
A list of components describing hurdle models for predicting the numbers of
trips by walking, biking, and taking public transportation.
}
\keyword{datasets}
