% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyDvmtReductions.R
\docType{data}
\name{ApplyDvmtReductionsSpecifications}
\alias{ApplyDvmtReductionsSpecifications}
\title{Specifications list for the ApplyDvmtReductions module.}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
ApplyDvmtReductions.R script.
}
\usage{
ApplyDvmtReductionsSpecifications
}
\description{
A list containing specifications for the ApplyDvmtReductions module.
}
\keyword{datasets}
