% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{assignEv}
\alias{assignEv}
\title{Function to identify electric vehicles}
\usage{
assignEv(Hh_df, Veh_df, EvRangePropYr_df, CurrYear,
  UseMaxDvmtCriterion = FALSE)
}
\arguments{
\item{Hh_df}{A household data frame consisting of variables required for calculation.}

\item{Veh_df}{A vehicle data frame consisting of variables required for calculation.}

\item{EvRangePropYr_df}{A data frame consisting of expected range of EV.}

\item{UseMaxDvmtCriterion}{A logical to indicated whether to use max dvmt criteria.}
}
\value{
A list of identifying the powertrain, dvmt, and efficiency of vehicles by
powertrain.
}
\description{
\code{assignEv} identifies electric vehicles.
}
\details{
This function takes a data frame of households and vehicles, and a data frame consisting
of expected range of electric vehicles to identify purely electric vehicles from
HEV/PHEV.
}
