% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{assignPHEV}
\alias{assignPHEV}
\title{Function to identify HEV/PHEV}
\usage{
assignPHEV(Hh_df, Veh_df, PhevRangePropYr_df, CurrYear, PhevPropModel_ls,
  HevMpgPropYr_df, OptimProp = 0)
}
\arguments{
\item{Hh_df}{A household data frame consisting of variables required for calculation.}

\item{Veh_df}{A vehicle data frame consisting of variables required for calculation.}

\item{PhevRangePropYr_df}{A data frame consisting of expected range of PHEV.}

\item{CurrYear}{The year for which the assignment of powertrian should be done.}

\item{PhevPropModel_ls}{A list of PHEV proportional models}

\item{HevMpgPropYr_df}{A data frame consisting of expectedrange of HEV.}

\item{OptimProp}{A numeric value indicating the proportion of vehicles that need
to be optimized.}
}
\value{
A list of identifying the powertrain, dvmt, and efficiency of vehicles by
powertrain.
}
\description{
\code{assignPHEV} identifies HEV/PHEV.
}
\details{
This function takes a data frame of households and vehicles, a list of
proportional model for identiying PHEVs, and a data frame of expected
range of HEVs and PHEVs to identify the powertrains of vehicles.
}
