% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleDvmt.R
\name{AssignVehicleDvmt}
\alias{AssignVehicleDvmt}
\title{Assign household DVMT to household vehicles.}
\usage{
AssignVehicleDvmt(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleDvmt} assigns DVMT to each household vehicle.
}
\details{
This function assigns household DVMT to each vehicle as a function of
household DVMT and precalculated proportional allocation of DVMT to vehicles.
}
