% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleDvmt.R
\docType{data}
\name{AssignVehicleDvmtSpecifications}
\alias{AssignVehicleDvmtSpecifications}
\title{Specifications list for AssignVehicleDvmt module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleDvmt.R script.
}
\usage{
AssignVehicleDvmtSpecifications
}
\description{
A list containing specifications for the AssignVehicleDvmt module.
}
\keyword{datasets}
