% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleDvmtSplit.R
\name{AssignVehicleDvmtSplit}
\alias{AssignVehicleDvmtSplit}
\title{Calculate the proportion of household VMT carried by each household vehicle.}
\usage{
AssignVehicleDvmtSplit(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleDvmtSplit} calculate proportions of household DVMT
allocated to each household.
}
\details{
This function calculates how household VMT is split among household
vehicles. The proportion of household VMT allocated to each vehicle is
calculated.
}
