% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleDvmtSplit.R
\docType{data}
\name{AssignVehicleDvmtSplitSpecifications}
\alias{AssignVehicleDvmtSplitSpecifications}
\title{Specifications list for AssignVehicleDvmtSplit module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleDvmtSplit.R script.
}
\usage{
AssignVehicleDvmtSplitSpecifications
}
\description{
A list containing specifications for the AssignVehicleDvmtSplit module.
}
\keyword{datasets}
