% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\docType{data}
\name{AveSovPropModels_ls}
\alias{AveSovPropModels_ls}
\title{SOV travel proportion model}
\format{A list having the following components:
\describe{
  \item{Model}{A list of equations to estimate SOV DVMT with 2, 5, 10, 15, and 20 mile thresholds.}
  \item{Parm}{Contains the alpha and beta parameters for the logistic transformation of model equations}
}}
\source{
CalculatePolicyVmt.R script.
}
\usage{
AveSovPropModels_ls
}
\description{
A list containing the SOV DVMT proportion model equation and relevant parameters.
}
\keyword{datasets}
