% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\name{calcAveSovProp}
\alias{calcAveSovProp}
\title{Function to calculate proportion of single occupant vehicles}
\usage{
calcAveSovProp(Data__, AveSovPropModels_, Threshold)
}
\arguments{
\item{Data__}{A household data frame consisting of household attributes used to
calculate proportion of single occupant vehicles.}

\item{AveSovPropModels_}{A list of SOV propotion models.}

\item{Threshold}{A numeric (miles) for which SOV model should be used}
}
\value{
An array of integers representing the number of single occupant vehicles
}
\description{
\code{calcAveSovProp} calculates proportion of single occupant vehicles.
}
\details{
This function takes a data frame of households, a distance threshold, and
a list of models which are used to calculate proportion of single
occupant vehicles for each household.
}
