% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\name{calcParkCostAdj}
\alias{calcParkCostAdj}
\title{Function to calculate adjusted parking costs for households}
\usage{
calcParkCostAdj(Data__, Park_)
}
\arguments{
\item{Data__}{A household data frame consisting of household attributes used to
calculate adjusted parking costs for households.}

\item{Park_}{A list for the attributes of households that uses parking.}
}
\value{
A list containing the daily parking cost, cash-out income adjustment
and parking cost per mile for each household.
}
\description{
\code{calcParkCostAdj} calculates adjusted parking costs for households.
}
\details{
This function takes a data frame of households and a list containing
attributes of the households that uses parking to calculate adjusted
parking costs for those households.
}
