% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{calculateAveElectricityCo2e}
\alias{calculateAveElectricityCo2e}
\title{Function to calculate average Co2 equivalent gas emissions for electricity
consumed by vehicle type}
\usage{
calculateAveElectricityCo2e(ForecastYear = NULL, PowerCo2Ft = NULL,
  OutputType = "MetricTons")
}
\arguments{
\item{ForecastYear}{An integer indicating the forecast year.}

\item{PowerCo2Ft}{A data frame containing the intensity of carbon by fuel types and electricity.}

\item{OutputType}{A string indicating the units of the output. ("MetricTons":Default or "Pounds")}
}
\value{
A named array indicating the average Co2 equivalent gas emissions by vehicle type.
}
\description{
\code{calculateAveElectricityCo2e} calculates average Co2 equivalent gas emissions
by vehicle and fuel type.
}
\details{
This function uses the composition of fuel and the proportion of fuel used
by vehicle type to calculate the average Co2 equivalent gas emissions for
a specific forecast year.
}
