% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{calculateAveFuelCo2e}
\alias{calculateAveFuelCo2e}
\title{Function to calculate average Co2 equivalent gas emissions for fuels by
vehicle type}
\usage{
calculateAveFuelCo2e(ForecastYear = NULL, FuelProp = NULL,
  FuelComp = NULL, FuelCo2Ft = NULL, MJPerGallon = 121,
  OutputType = "MetricTons")
}
\arguments{
\item{ForecastYear}{An integer indicating the forecast year.}

\item{FuelProp}{A data frame consisting of fuel proportion used by vehicle types.}

\item{FuelComp}{A data frame consisting of composition of fuels.}

\item{FuelCo2Ft}{A data frame containing the intensity of carbon by fuel types.}

\item{MJPerGallon}{A numeric indicating the energy per gallon of fuel. (Default: 121)}

\item{OutputType}{A string indicating the units of the output. ("MetricTons":Default or "Pounds")}
}
\value{
A named array indicating the average Co2 equivalent gas emissions by vehicle type.
}
\description{
\code{calculateAveFuelCo2e} calculates average Co2 equivalent gas emissions
by vehicle and fuel type.
}
\details{
This function uses the composition of fuel and the proportion of fuel used
by vehicle type to calculate the average Co2 equivalent gas emissions for
a specific forecast year.
}
