% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateInducedDemand.R
\name{CalculateInducedDemand}
\alias{CalculateInducedDemand}
\title{Calculate various attributes of travel demands for each household
and vehicle using future data}
\usage{
CalculateInducedDemand(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateInducedDemand} calculate various attributes of travel
demands for each household and vehicle using future data
}
\details{
This function calculates dvmt by placetypes, households, and vehicles.
It also calculates fuel gallons consumed, total fuel cost, and Co2 equivalent
gas emission for each household using future data.
}
