% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\docType{data}
\name{CalculatePolicyVmtSpecifications}
\alias{CalculatePolicyVmtSpecifications}
\title{Specifications list for CalculatePolicyVmt module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculatePolicyVmt.R script.
}
\usage{
CalculatePolicyVmtSpecifications
}
\description{
A list containing specifications for the CalculatePolicyVmt module.
}
\keyword{datasets}
