% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{CalculateTravelDemand}
\alias{CalculateTravelDemand}
\title{Calculate various attributes of travel demands for each household
and vehicle.}
\usage{
CalculateTravelDemand(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateTravelDemand} calculate various attributes of travel
demands for each household and vehicle.
}
\details{
This function calculates dvmt by placetypes, households, and vehicles.
It also calculates fuel gallons consumed, total fuel cost, and Co2 equivalent
gas emission for each household.
}
