% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\docType{data}
\name{CalculateTravelDemandSpecifications}
\alias{CalculateTravelDemandSpecifications}
\title{Specifications list for CalculateTravelDemand module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateTravelDemand.R script.
}
\usage{
CalculateTravelDemandSpecifications
}
\description{
A list containing specifications for the CalculateTravelDemand module.
}
\keyword{datasets}
