% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R,
%   R/CalculateTravelDemandFuture.R
\name{calculateVehFuelCo2}
\alias{calculateVehFuelCo2}
\alias{calculateVehFuelCo2}
\title{Function to calculate average Co2 equivalent gas emissions for fuels for
the households and the vehicles}
\usage{
calculateVehFuelCo2(Hh_df, Vehicles_df, AveFuelCo2e)

calculateVehFuelCo2(Hh_df, Vehicles_df, AveFuelCo2e)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes used to
assign Co2 equivalent emissions.}

\item{Vehicles_df}{A vehicle data frame consisting of vehicle attributes used to
assign Co2 equivalent emissions.}

\item{AveFuelCo2e}{A named array indicating the average Co2 equivalent gas emissions
by vehicle type.}

\item{Hh_df}{A household data frame consisting of household attributes used to
assign Co2 equivalent emissions.}

\item{Vehicles_df}{A vehicle data frame consisting of vehicle attributes used to
assign Co2 equivalent emissions.}

\item{AveFuelCo2e}{A named array indicating the average Co2 equivalent gas emissions
by vehicle type.}
}
\value{
A list containing assignment of gas emissions.

A list containing assignment of gas emissions.
}
\description{
\code{calculateVehFuelCo2} calculates average Co2 equivalent gas emissions
for the households and the vehicles.

\code{calculateVehFuelCo2} calculates average Co2 equivalent gas emissions
for the households and the vehicles.
}
\details{
This function uses Dvmt and fuel efficiency of the vehicles, along with
the average Co2 equivalent gas emissions by vehicle type to assign
emissions to the vehicles.

This function uses Dvmt and fuel efficiency of the vehicles, along with
the average Co2 equivalent gas emissions by vehicle type to assign
emissions to the vehicles.
}
