% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{calculateVehFuelElectricCo2}
\alias{calculateVehFuelElectricCo2}
\title{Function to calculate average Co2 equivalent gas emissions for fuels and
electricity for the households and the vehicles}
\usage{
calculateVehFuelElectricCo2(Hh_df, Vehicles_df, AveFuelCo2e,
  AveElectricCo2e)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes used to
assign Co2 equivalent emissions.}

\item{Vehicles_df}{A vehicle data frame consisting of vehicle attributes used to
assign Co2 equivalent emissions.}

\item{AveFuelCo2e}{A named array indicating the average Co2 equivalent gas emissions
by vehicle type per gallon of fuel.}

\item{AveElectricCo2e}{A numeric indicating the average Co2 equivalent gas emissions
per Kwh of electricity.}
}
\value{
A list containing assignment of gas emissions.
}
\description{
\code{calculateVehFuelElectricCo2} calculates average Co2 equivalent gas emissions
for the households and the vehicles.
}
\details{
This function uses Dvmt and fuel efficiency of the vehicles, along with
the average Co2 equivalent gas emissions by vehicle type to assign
emissions to the vehicles.
}
