% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleEnergySplit.R
\name{CalculateVehicleEnergySplit}
\alias{CalculateVehicleEnergySplit}
\title{Calculate energy type use proportions of household vehicles.}
\usage{
CalculateVehicleEnergySplit(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateVehicleEnergySplit} calculates the energy type (electric vs.
hydrocarbon) use proportion of household vehicles.
}
\details{
This function calculates the proportion of DVMT of each household vehicle
that is powered by on-board hydrocarbon fuel such as gasoline and the portion
that is powered by electricity stored in the vehicle's battery.
}
