% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleEnergySplit.R
\docType{data}
\name{CalculateVehicleEnergySplitSpecifications}
\alias{CalculateVehicleEnergySplitSpecifications}
\title{Specifications list for CalculateVehicleEnergySplit module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateVehicleEnergySplit.R script.
}
\usage{
CalculateVehicleEnergySplitSpecifications
}
\description{
A list containing specifications for the CalculateVehicleEnergySplit module.
}
\keyword{datasets}
