% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleTrips.R
\name{CalculateVehicleTrips}
\alias{CalculateVehicleTrips}
\title{Main module function to calculate household vehicle trips}
\usage{
CalculateVehicleTrips(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateVehicleTrips} calculates vehicle trips for each household.
}
\details{
This function calculates vehicle trips for households.
}
