% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivertSovTravel.R
\name{DivertSovTravel}
\alias{DivertSovTravel}
\title{Assign household DVMT to household vehicles.}
\usage{
DivertSovTravel(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{DivertSovTravel} calculates the proportion of household DVMT that is
diverted to meet Azone SOV diversion goals.
}
\details{
This function calculates the proportion of household DVMT that is diverted to
meet Azone SOV diversion goals. It calculates the amount of SOV travel
of each households that is in tours having round-trip distances of 20 miles
or less. Based on that calculation, it calculates the total DVMT to be
diverted. It then allocates the DVMT to be diverted as a function of the
household's alternative mode trip making, assuming that households that are
predicted to make more alternative mode trips would be more likely to make
additional diversions. Finally, it calculates the proportion of DVMT diverted
for each household.
}
