% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivertSovTravel.R
\docType{data}
\name{DivertSovTravelSpecifications}
\alias{DivertSovTravelSpecifications}
\title{Specifications list for DivertSovTravel module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
DivertSovTravel.R script.
}
\usage{
DivertSovTravelSpecifications
}
\description{
A list containing specifications for the DivertSovTravel module.
}
\keyword{datasets}
