% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\name{idPayingParkers}
\alias{idPayingParkers}
\title{Function to identify households that pays for parking}
\usage{
idPayingParkers(Data__, PropWrkPkg, PropWrkChrgd, PropCashOut,
  PropOthChrgd, LabForcePartRate = 0.65, PkgCost, PropWrkTrav = 0.22,
  WrkDaysPerYear = 260)
}
\arguments{
\item{Data__}{A household data frame consisting of household attributes used to
identify households that pays for parking.}

\item{PropWrkPkg}{A numeric for proportion of employees that park at work}

\item{PropWrkChrgd}{A numeric for proportion of employers that charge for parking}

\item{PropCashOut}{A numeric for employment parking that is converted from being
free to pay under a “cash-out buy-back” type of program}

\item{PropOthChrgd}{A numeric for proportion of other parking that is not free}

\item{LabForcePartRate}{A numeric indicating the labor force participation rate
(Default: 0.65)}

\item{PkgCost}{A numeric for average daily parking cost}

\item{PropWrkTrav}{A numeric for proportion of employees that travel to work}

\item{WrkDaysPerYear}{A numeric for number of working days per year}
}
\value{
A list.
}
\description{
\code{idPayingParkers} identifies households that pays for parking.
}
\details{
This function takes a data frame of households and various parameters
to identify households that pays for parking.
}
