% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{Initialize}
\alias{Initialize}
\title{Check and optional roadway base year DVMT parameters for consistency.}
\usage{
Initialize(L)
}
\arguments{
\item{L}{A list containing data from preprocessing supplied optional input
files returned by the processModuleInputs function. This list has two
components: Errors and Data.}
}
\value{
A list that is the same as the input list with an additional
Warnings component.
}
\description{
\code{Initialize} checks optional roadway base year DVMT parameters for
consistency and returns those that have values (i.e. not NA). Errors are
returned for inconsistent values.
}
\details{
This function processes optional user roadway base year DVMT inputs to check
that values are consistent. Errors are returned for inconsistent values.
The script checks whether proportions data that should sum to 1 does, whether
urbanized area lookup names are in the urbanized area table, and whether
base year DVMT data have consistent values.
}
