% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePolicyVmt.R
\docType{data}
\name{LtVehOwnModels_ls}
\alias{LtVehOwnModels_ls}
\title{Light Vehicle ownership model}
\format{A list having the following components:
\describe{
  \item{Metro}{Light vehicle ownership model for the metropolitan region}
  \item{NonMetro}{Light vehicle ownership model for the non-metropolitan region}
}}
\source{
AssignVehicleFeatures.R script.
}
\usage{
LtVehOwnModels_ls
}
\description{
A list containing the light vehicle ownership model equation and other information
needed to implement the vehicle ownership model.
}
\keyword{datasets}
