% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{predictAveDvmt}
\alias{predictAveDvmt}
\title{Function to predict average DVMT}
\usage{
predictAveDvmt(Hh_df, Model_ls, Type)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes used to
predict average DVMT.}

\item{Model_ls}{A list of DVMT assignment models.}

\item{Type}{A string indicating the region type. ("Metro": Default, or "NonMetro")}
}
\value{
A matrix containing average, maximum, and 95th percentile of daily
vehicle miles traveled.
}
\description{
\code{predictAveDvmt} predicts average DVMT for each household.
}
\details{
This function takes a data frame of households and a list of models which
are used to predict the average daily vehicle miles traveled for each
household.
}
