% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{predictMaxDvmt}
\alias{predictMaxDvmt}
\title{Function to predict max and 95th percentile DVMT}
\usage{
predictMaxDvmt(Hh_df, Model_ls, Type)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes used to
predict average DVMT.}

\item{Model_ls}{A list of DVMT assignment models.}

\item{Type}{A string indicating the region type. ("Metro": Default, or "NonMetro")}
}
\value{
A matrix containing maximum and 95th percentile of daily
vehicle miles traveled.
}
\description{
\code{predictMaxDvmt} predicts max and 95th percentile DVMT for each household.
}
\details{
This function takes a data frame of households and a list of models which
are used to predict the max and 95th percentile daily vehicle miles traveled for each
household.
}
