% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivertSovTravel.R
\docType{data}
\name{SovModel_ls}
\alias{SovModel_ls}
\title{Model of proportion of household average DVMT in single-occupant vehicle
tours having round-trip distances of 20 miles or less.}
\format{A list having two components:
Pow: the power transform of the SOV DVMT proportion;
Prop: a string representation of the model formula;}
\source{
DivertSovTravel.R
}
\usage{
SovModel_ls
}
\description{
A list having two components: Pow - the power transform of the SOV DVMT
proportion, and Prop - a string representation of the model formula.
}
\keyword{datasets}
