% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleTrips.R
\docType{data}
\name{VehTripModels_ls}
\alias{VehTripModels_ls}
\title{Household vehicle trip models}
\format{A list having 'Metro' and 'NonMetro' components. Each component
contains a string representing a hurdle model for computing household trips;}
\source{
CalculateVehicleTrips.R script.
}
\usage{
VehTripModels_ls
}
\description{
A list of components describing hurdle models for predicting the numbers of
household vehicle trips.
}
\keyword{datasets}
