% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{calculateCosts}
\alias{calculateCosts}
\title{Function to calculate total fuel cost per mile for the households}
\usage{
calculateCosts(Hh_df, Costs, NonPrivateFactor = 5)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes used to
caolculate total fuel cost per mile.}

\item{Costs}{A named numeric consisting of the costs, and/or tax  information.}

\item{NonPrivateFactor}{A numeric.}
}
\value{
A list containing various fuel costs.
}
\description{
\code{calculateCosts} calculates total fuel cost per mile for the
households.
}
\details{
This function uses fuel cost, gas tax, carbon cost, and vehicle miles traveled
by vehicles of the households to calculate total fuel cost per mile.
}
