% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTravelDemand.R
\name{calculateVehDvmt}
\alias{calculateVehDvmt}
\title{Function to calculate DVMT for all vehicles}
\usage{
calculateVehDvmt(Hh_df, Vehicles_df)
}
\arguments{
\item{Hh_df}{A household data frame consisting average DVMT and household id.}

\item{Vehicles_df}{A vehicle data frame consisting of variables used for DVMT assignment.}
}
\value{
A numeric vector of DVMT.
}
\description{
\code{calculateVehDvmt} calculates daily vehicle miles traveled for each vehicle in
a household.
}
\details{
This function takes a data frame of households and vehicles, and assigns the household
average DVMT to the vehicles.
}
