% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivertSovTravel.R
\docType{data}
\name{SovModel_ls}
\alias{SovModel_ls}
\title{Models of the proportion of household average DVMT in single-occupant vehicle
tours having round-trip distances of 20 miles or less, and of the average
trip length in such tours}
\format{A list having 6 components:
ZeroSovModel: summary statistics for the model to predict the likelihood that
a household has no qualifying tours on the survey day;
SovDvmtModel: summary statistics for the model to predict miles of travel in
qualifying tours if there is some travel in qualifying tours;
SovPropModel: a list that describes and documents a linear model for
predicting the proportion of average DVMT in qualifying tours;
MetroTrpLenModel: summary statistics for the model to predict average SOV
trip length for metropolitan households;
NonMetroTrpLenModel: summary statistics for the model to predict average SOV
trip length for non-metropolitan households;
SovTrpLenModel: a list composed of two lists (Metro and NonMetro) which each
describe a linear model for predicting average SOV trip length for trips in
qualifying tours.}
\source{
DivertSovTravel.R
}
\usage{
SovModel_ls
}
\description{
A list having components to describe or document the model for predicting the
proportion of household DVMT in tours having round-trip distances of 20 miles
or less, and the model of the average length of trips in such tours.
}
\keyword{datasets}
